Added tracing example for Agents
