+++ dotnet/src/InternalUtilities/samples/InternalUtilities/TestConfiguration.cs
@@ -48,6 +48,7 @@ public sealed class TestConfiguration
     public static GoogleAIConfig GoogleAI => LoadSection<GoogleAIConfig>();
     public static VertexAIConfig VertexAI => LoadSection<VertexAIConfig>();
     public static AzureCosmosDbMongoDbConfig AzureCosmosDbMongoDb => LoadSection<AzureCosmosDbMongoDbConfig>();
+    public static ApplicationInsightsConfig ApplicationInsights => LoadSection<ApplicationInsightsConfig>();
 
     private static T LoadSection<T>([CallerMemberName] string? caller = null)
     {
@@ -261,6 +262,11 @@ public sealed class TestConfiguration
     {
         public string ConnectionString { get; set; }
         public string DatabaseName { get; set; }
+    }
+
+    public class ApplicationInsightsConfig
+    {
+        public string ConnectionString { get; set; }
     }
 
     /// <summary>
